hand-hygiene-compliance-with-azure-percept/modules/evaluatemodule/requirements.txt
azure-iot-device~=2.7.0
